# native: Use asm.Dscal in Dscal