--- a/native/level1double.go
+++ b/native/level1double.go
@@ -578,16 +578,14 @@ func (Implementation) Dscal(n int, alpha float64, x []float64, incX int) {
 		}
 		return
 	}
-	if incX > 0 && (n-1)*incX >= len(x) {
+	if (n-1)*incX >= len(x) {
 		panic(badX)
 	}
 	if n < 1 {
 		if n == 0 {
 			return
 		}
-		if n < 1 {
-			panic(negativeN)
-		}
+		panic(negativeN)
 	}
 	if alpha == 0 {
 		if incX == 1 {
@@ -595,10 +593,12 @@ func (Implementation) Dscal(n int, alpha float64, x []float64, incX int) {
 			for i := range x {
 				x[i] = 0
 			}
+			return
 		}
 		for ix := 0; ix < n*incX; ix += incX {
 			x[ix] = 0
 		}
+		return
 	}
 	if incX == 1 {
 		x = x[:n]
@@ -610,5 +610,4 @@ func (Implementation) Dscal(n int, alpha float64, x []float64, incX int) {
 	for ix := 0; ix < n*incX; ix += incX {
 		x[ix] *= alpha
 	}
-	return
 }