--- a/native/level1single.go
+++ b/native/level1single.go
@@ -602,16 +602,14 @@ func (Implementation) Sscal(n int, alpha float32, x []float32, incX int) {
 		}
 		return
 	}
-	if incX > 0 && (n-1)*incX >= len(x) {
+	if (n-1)*incX >= len(x) {
 		panic(badX)
 	}
 	if n < 1 {
 		if n == 0 {
 			return
 		}
-		if n < 1 {
-			panic(negativeN)
-		}
+		panic(negativeN)
 	}
 	if alpha == 0 {
 		if incX == 1 {
@@ -619,20 +617,18 @@ func (Implementation) Sscal(n int, alpha float32, x []float32, incX int) {
 			for i := range x {
 				x[i] = 0
 			}
+			return
 		}
 		for ix := 0; ix < n*incX; ix += incX {
 			x[ix] = 0
 		}
+		return
 	}
 	if incX == 1 {
-		x = x[:n]
-		for i := range x {
-			x[i] *= alpha
-		}
+		asm.SscalUnitary(alpha, x[:n])
 		return
 	}
 	for ix := 0; ix < n*incX; ix += incX {
 		x[ix] *= alpha
 	}
-	return
 }