I wonder if we shouldn't keep the PosInc case as Go loop. Also, mat64.Vector does not allow negative inc, BLAS does nothing when negative inc, so we could safely drop the last argument from the assembler routines.
```
% benchstat before.txt after.txt 
name                   old time/op  new time/op  delta
DscalSmallUnitaryInc   12.6ns ± 0%   9.4ns ± 0%  -25.26%  (p=0.029 n=4+4)
DscalSmallPosInc       13.4ns ± 0%  13.2ns ± 0%   -1.49%  (p=0.029 n=4+4)
DscalMediumUnitaryInc   690ns ± 0%   164ns ± 0%  -76.25%  (p=0.029 n=4+4)
DscalMediumPosInc      1.14µs ± 0%  1.15µs ± 0%   +1.36%  (p=0.029 n=4+4)
DscalLargeUnitaryInc   71.1µs ± 0%  42.8µs ± 0%  -39.87%  (p=0.029 n=4+4)
DscalLargePosInc        197µs ± 0%   183µs ± 1%   -7.44%  (p=0.029 n=4+4)
DscalHugeUnitaryInc    9.92ms ± 1%  8.00ms ± 1%  -19.38%  (p=0.029 n=4+4)
DscalHugePosInc        41.4ms ± 0%  41.1ms ± 0%   -0.76%  (p=0.029 n=4+4)
```