@kortschak 2016-01-15
> I wonder if we shouldn't keep the PosInc case as Go loop.

You mean [here](https://github.com/gonum/blas/pull/150/files#diff-876d7c9ede015498609b19445fd534d3L603)?

> Also, mat64.Vector does not allow negative inc, BLAS does nothing when negative inc, so we could safely drop the last argument from the assembler routines.

I don't understand this.