@kortschak 2016-01-15
Clear now thanks.

Both those SGTM. The second should add a comment that increments are always positive - we use them that way now, but we need it documented at the site.