@vladimir-ch 2016-01-15
Commit dropped. PTAL @kortschak